+++ spec/iparty/max_mind/singleton_spec.rb
@@ -7,10 +7,20 @@ RSpec.describe IParty::MaxMind do
   after { IParty.config.singletons = false }
 
   context "with global singletons" do
-    it "populates" do
-      IParty.config.singletons = true
+    it "eagerly populates" do
+      IParty.config.init_singletons!
+      expect(IParty.config.singletons.keys).to eq %w[GeoLite2-ASN GeoLite2-Country GeoLite2-City]
+    end
 
-      expect(IParty.config.singletons).to be_a Hash
+    it "lazily populates" do
+      IParty.config.singletons = {}
+      IParty::MaxMind.db(:ASN)
+      expect(IParty.config.singletons.keys).to eq %w[GeoLite2-ASN]
+    end
+
+    it "auto populates on first use" do
+      IParty.config.singletons = true
+      IParty::MaxMind.db(:ASN)
       expect(IParty.config.singletons.keys).to eq %w[GeoLite2-ASN GeoLite2-Country GeoLite2-City]
     end
 
@@ -18,7 +28,7 @@ RSpec.describe IParty::MaxMind do
       ghash = {}
       IParty.config.singletons = -> { ghash }
       IParty.config.init_singletons!
-      expect(IParty.config.singletons.call.keys).to eq %w[GeoLite2-ASN GeoLite2-Country GeoLite2-City]
+      expect(ghash.keys).to eq %w[GeoLite2-ASN GeoLite2-Country GeoLite2-City]
     end
   end
 end